app/requirements.txt
pandas
numpy
Flask
Flask-Cors
flask-restx
mysql-connector-python
torch
transformers
tweet-preprocessor